refactoring numbers view and downloads
