+++ api/endpoints/models/download_models.py
@@ -1,0 +1,16 @@
+from pydantic import BaseModel
+from enum import Enum
+from .general_models import GeneralInput
+
+
+class DownloadData(str, Enum):
+    table = "table"
+    numbers = "numbers"
+
+
+class DownloadInput(GeneralInput):
+    download_data: DownloadData = DownloadData.table
+
+
+class DownloadOutput(BaseModel):
+    __root__: bytes
