+++ api/main.py
@@ -16,6 +16,7 @@ from .endpoints import (
     menu,
     utils,
     links,
+    download,
 )
 
 API_PREFIX = "/api-db" if os.environ["PROD"] == "1" else "/api-db"
@@ -28,6 +29,7 @@ APP.add_middleware(
 
 APP.include_router(search.router)
 APP.include_router(graph_view.router)
+APP.include_router(download.router)
 APP.include_router(table_view.router, prefix="/table-view")
 APP.include_router(text_view.router, prefix="/text-view")
 APP.include_router(numbers_view.router, prefix="/numbers-view")
